Fehler behoben

+++ Raytracer/rt/materials/phong.cpp
@@ -8,6 +8,7 @@
 #include <rt/materials/phong.h>
 #include <rt/materials/material.h>
 #include <rt/textures/texture.h>
+#include <math.h>
 
 namespace rt {
 
@@ -18,7 +19,7 @@ PhongMaterial::PhongMaterial(Texture* specular, float exponent): specular(specul
 RGBColor PhongMaterial::getReflectance(const Point& texPoint, const Vector& normal, const Vector& outDir,
 		const Vector& inDir) const {
 	Vector RI = 2*dot(inDir, normal)*normal - inDir;
-	float refl = pow(dot(RI, outDir), exponent) / dot(inDir, normal);
+	float refl = powf(dot(RI, outDir), exponent) / dot(inDir, normal);
 	return (specular->getColor(texPoint))*refl;
 }
